backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
psycopg2-binary
pytest
torch
torchvision
torchaudio
opencv-python
librosa
reportlab
python-gnupg
pycryptodome
pillow
numpy
ffmpeg-python
